Create git-daemon container to serve pipeline as code repository
I pivoted to creating an environment in which your pipeline code is in its own repository. For the GoCD server to access this properly I had to make sure the code was served. The easiest way to do this was to add bankiru's git-daemon image to the docker-compose and volume mount the pipeline configuration.