Externalize the pipeline code repository
Made the docker-compose configuration for the git-daemon take an environment variable named PIPELINE_CODE for the location of the pipeline as code repository. This will volume mount the pipeline code, thereby serving the correct pipeline code for the GoCD server to use. This allows consumers to utilize the project without having to change the source code.