Allow go-agents to run docker
The go-agent service has the docker.sock volume mounted to connect to the docker daemon and the go-agent Dockerfile is installing the docker engine. This is necessary to be able to run docker commands inside of the go-agent. Mason and I discovered that "chmod +s $(which docker)" will run the docker engine as root whenever it is invoked. Since there appears to be a permissioning issue with the go user being able to access the docker daemon, this appears to be a good workaround for the time being.