Mount dockerhub credentials
The dockerhub credentials are mounted for push and pull purposes through the lifecycle of our pipelines. The pipeline code contains a task at the beginning of each job which logs into docker just in case it is a new agent. Eventually, this should be automated into the creation of the agent on startup.